fixed MX priority, added SRV
git-svn-id: http://svn.code.sf.net/p/dnsjava/code/trunk@34 c76caeb1-94fd-44dd-870f-0c9d92034fc1

+++ org/xbill/DNS/dnsMXRecord.java
@@ -3,7 +3,7 @@ import java.io.*;
 
 public class dnsMXRecord extends dnsRecord {
 
-short priority;
+int priority;
 dnsName name;
 
 public dnsMXRecord(dnsName rname, short rclass) {
@@ -21,7 +21,7 @@ public dnsMXRecord(dnsName rname, short rclass, int rttl, int priority,
 }
 
 void parse(CountedDataInputStream in, dnsCompression c) throws IOException {
-	priority = (short)in.readUnsignedShort();
+	priority = in.readUnsignedShort();
 	name = new dnsName(in, c);
 }
